requirements.txt
numpy
Pillow
vtk
scipy